fix:将资讯卡片组件放入Home组件中（组件形式无法拖拽布局）
