--- src/pages/Home/all.json
@@ -1,81 +1,0 @@
-
-{
-    "tab": "全部",
-    "pageInfo": [
-        {
-            "x": 1,
-            "y": 1,
-            "w": 3,
-            "h": 4,
-            "title":"知乎热榜",
-            "key":"zhihu",
-            "url":"www.zhihu.com",
-            "content":[{
-                "itemTitle": "热榜1",
-                "itemContent": "热榜内容1"
-            },
-            {
-                "itemTitle": "热榜2",
-                "itemContent": "热榜内容2"
-            },
-            {
-                "itemTitle": "热榜3",
-                "itemContent": "热榜内容3"
-            },
-            {
-                "itemTitle": "热榜4",
-                "itemContent": "热榜内容4"
-            }]
-        },
-        {
-            "x": 2,
-            "y": 2,
-            "w": 3,
-            "h": 4,
-            "title":"微博热榜",
-            "key":"weibo",
-            "url":"www.weibo.com",
-            "content":[{
-                "itemTitle": "热榜1",
-                "itemContent": "热榜内容1"
-            },
-            {
-                "itemTitle": "热榜2",
-                "itemContent": "热榜内容2"
-            },
-            {
-                "itemTitle": "热榜3",
-                "itemContent": "热榜内容3"
-            },
-            {
-                "itemTitle": "热榜4",
-                "itemContent": "热榜内容4"
-            }]
-        },
-        {
-            "x": 3,
-            "y": 1,
-            "w": 2,
-            "h": 4,
-            "title":"虎扑",
-            "key":"hupu",
-            "url":"www.hupu.com",
-            "content":[{
-                "itemTitle": "热榜1",
-                "itemContent": "热榜内容1"
-            },
-            {
-                "itemTitle": "热榜2",
-                "itemContent": "热榜内容2"
-            },
-            {
-                "itemTitle": "热榜3",
-                "itemContent": "热榜内容3"
-            },
-            {
-                "itemTitle": "热榜4",
-                "itemContent": "热榜内容4"
-            }]
-        }
-    ]
-}
